Update create-your-first-app.mdx
this command doesn't exist, and I added it here to help troubleshoot the error easily. people sometime need more options when they are coding so  i think this command will be usefull I added the following text to assist in troubleshooting the error because the command in question does not currently exist. It's important to have more options available when coding, so I believe this command could be quite useful.